+++ Sheets_Manager/module/sheets_manager.py
@@ -1,5 +1,5 @@
 import gspread
-from Files_Handler.file_handler import Files_Handling
+from Files_Handler.module.file_handler import Files_Handling
  
 class Sheets_Manager(Files_Handling):
